Refactoring and ReadMe
